requirements.txt
openai
langchain
chromadb
tiktoken
streamlit
python-dotenv
GitPython
sentence-transformers